requirements.txt
psycopg2-binary==2.9.9
requests==2.31.0
pandas==2.1.4
openpyxl==3.1.2
python-dotenv==1.0.0

# Testing
pytest==7.4.3
pytest-cov==4.1.0
coverage==7.3.2 